refactor(Total Balance): Remove unused files
