--- packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/MenuTop/Balance/template.error.js
@@ -1,21 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-
-import { Text } from 'blockchain-info-components'
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 5px;
-  box-sizing: border-box;
-`
-
-export default (props) => (
-  <Wrapper>
-    <Text size='12px' weight={300} color='red'>
-      {props.children}
-    </Text>
-  </Wrapper>
-)
