+++ packages/blockchain-wallet-v4-frontend/src/layouts/Wallet/MenuTop/Balance/template.js
